pyproject.toml
[project]
name = "infinity-bot-v2-simplize"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "machinacanis", email = "hf4minecraft@163.com"},
]
dependencies = ["httpx>=0.28.1", "nonebot2[fastapi]>=2.4.1", "motor>=3.7.0", "nonebot-adapter-onebot>=2.4.6", "nonebot-plugin-alconna>=0.55.1", "cachetools>=5.5.1", "psutil>=7.0.0", "aiohttp>=3.11.12", "pillow>=11.1.0"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
